Workaround
Due to the Heroku hack, we cannot connect Heroku to Github. Therefore, had to use the Heroku CLI and Git to commit changes to our deployment for testing. I updated the existing index page, added formatting, updated the register form, added a nav bar, and updated the database fields in the insert.php query.